// generated/src/app/api-models/claim-api-detail-profsvc-w.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ClaimApiDetailProfsvcW {
  claimIdNo: string;
  lineNumber: string;
  svcFromDate: string;
  svcToDate: string;
  renderingProviderId: string;
  diagCodePointer1: string;
  diagCodePointer2: string;
  diagCodePointer3: string;
  diagCodePointer4: string;
  placeOfSvc: string;
  typeOfSvcCode: string;
  hcpcsProcedureCode: string;
  hcpcsModifier1: string;
  hcpcsModifier2: string;
  hcpcsModifier3: string;
  hcpcsModifier4: string;
  lineCharges: string;
  unitsOfSvc: string;
  anesthesiaOxygenMin: string;
  primaryPaidAmount: string;
  status: string;
  epsdtFamilyPlanningInd: string;
  emergencyServicesInd: string;
  cobIndicator: string;
  reservedLocalUseDet: string;
  cobPatLiabCvrgAmt: string;
  fullCvrgAmt: string;
  ocAllowedAmt: string;
  adjudDate: string;
  epsdt: string;
  familyPlanningIndicator: string;
}